refactor(rds): decompose module_utils/rds.py into _rds/ private subpackage
Split the monolithic rds.py (790 lines) into a _rds/ private subpackage
following the established _s3/, _autoscaling/, _elbv2/ pattern:

- common.py: error classes, Boto3ClientMethod, method routing
- waiters.py: wait_for_*_status functions
- transformations.py: parameter formatting, IAM role comparison
- api.py: boto3 API wrappers, call_method, handle_errors
- tags.py: get_tags, ensure_tags

Rewrote rds.py as a thin facade re-exporting all public symbols.
All 13 consuming modules continue importing from rds.py unchanged.

Reorganized tests/unit/plugins/module_utils/test_rds.py into a rds/
subdirectory with per-submodule test files matching the s3/ test pattern.
